campaign interaction step UI perfomance optimization
