+++ src/components/CampaignInteractionStepsForm.jsx
@@ -1,54 +1,16 @@
 import type from "prop-types";
 import React from "react";
 import RaisedButton from "material-ui/RaisedButton";
-import IconButton from "material-ui/IconButton";
-import DeleteIcon from "material-ui/svg-icons/action/delete";
-import { Card, CardHeader, CardText } from "material-ui/Card";
-import theme from "../styles/theme";
 import CampaignFormSectionHeading from "./CampaignFormSectionHeading";
-import HelpIconOutline from "material-ui/svg-icons/action/help-outline";
-import Form from "react-formal";
-import GSForm from "./forms/GSForm";
-import yup from "yup";
 import { makeTree } from "../lib";
 import { dataTest } from "../lib/attributes";
-import { StyleSheet, css } from "aphrodite";
 
-const styleSheet = StyleSheet.create({
-  errorMessage: {
-    color: theme.colors.red
-  }
-});
-
-const styles = {
-  pullRight: {
-    float: "right",
-    position: "relative",
-    top: "10px",
-    icon: "pointer"
-  },
-
-  cardHeader: {
-    backgroundColor: theme.colors.veryLightGray
-  },
-
-  interactionStep: {
-    borderLeft: `5px solid ${theme.colors.green}`,
-    marginBottom: 24
-  },
-
-  answerContainer: {
-    marginLeft: "35px",
-    marginTop: "10px",
-    borderLeft: `3px dashed ${theme.colors.veryLightGray}`
-  }
-};
+import InteractionStep from "./CampaignInteractionStep";
 
 export default class CampaignInteractionStepsForm extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      focusedField: null,
       availableActionsLookup: props.availableActions.reduce(
         (lookup, action) => {
           const toReturn = {
@@ -72,25 +34,9 @@ export default class CampaignInteractionStepsForm extends React.Component {
               answerActionsData: "",
               isDeleted: false
             }
-          ],
-      displayAllSteps: false
+          ]
     };
   }
-  /*
-    this.state.displayAllSteps is used to cause only the root interaction
-    node to render when the component first mounts.  ComponentDidMount sets
-    displayAllSteps to true, which forces render to run again, this time
-    including all interaction steps. This cuts half the time required to
-    render the interaction steps after clicking the header to expand the
-    interaction steps card.
-
-    FUTURE: reevaluate this after React >=16 upgrade
-  */
-  componentDidMount = () => {
-    if (!this.state.displayAllSteps) {
-      this.setState({ displayAllSteps: true });
-    }
-  };
 
   onSave = async () => {
     const tweakedInteractionSteps = this.state.interactionSteps.map(is => {
@@ -117,11 +63,9 @@ export default class CampaignInteractionStepsForm extends React.Component {
 
   addStep(parentInteractionId) {
     return () => {
-      const newId =
-        "new" +
-        Math.random()
-          .toString(36)
-          .replace(/[^a-zA-Z1-9]+/g, "");
+      // keep it simple – and sortable
+      const newId = `new${Date.now()}`;
+
       this.setState({
         interactionSteps: [
           ...this.state.interactionSteps,
@@ -165,9 +109,9 @@ export default class CampaignInteractionStepsForm extends React.Component {
   bumpStep(id) {
     return () => {
       const step = this.state.interactionSteps.find(is => is.id === id);
-      var livingSiblings = [];
-      var otherRelatives = [];
-      for (let is of this.state.interactionSteps) {
+      const livingSiblings = [];
+      const otherRelatives = [];
+      for (const is of this.state.interactionSteps) {
         if (
           is.parentInteractionId !== step.parentInteractionId ||
           step.isDeleted
@@ -233,209 +177,8 @@ export default class CampaignInteractionStepsForm extends React.Component {
     });
   }
 
-  formSchema = yup.object({
-    script: yup.string(),
-    questionText: yup.string(),
-    answerOption: yup.string(),
-    answerActions: yup.string(),
-    answerActionsData: yup.string()
-  });
-
-  renderInteractionStep(interactionStep, availableActions, title = "Start") {
-    const answerActions =
-      interactionStep.answerActions &&
-      availableActions[interactionStep.answerActions];
-    let clientChoiceData;
-    let instructions;
-
-    if (answerActions) {
-      clientChoiceData = answerActions.clientChoiceData;
-      instructions = answerActions.instructions;
-    }
-
-    return (
-      <div>
-        {interactionStep.parentInteractionId ? (
-          <div>
-            <DeleteIcon
-              style={styles.pullRight}
-              onTouchTap={this.deleteStep(interactionStep.id).bind(this)}
-            />
-            <RaisedButton
-              label="Bump"
-              onTouchTap={this.bumpStep(interactionStep.id).bind(this)}
-            />
-            <RaisedButton
-              label="Top"
-              onTouchTap={this.topStep(interactionStep.id).bind(this)}
-            />
-            <RaisedButton
-              label="Bottom"
-              onTouchTap={this.bottomStep(interactionStep.id).bind(this)}
-            />
-          </div>
-        ) : (
-          ""
-        )}
-        <Card
-          style={styles.interactionStep}
-          ref={interactionStep.id}
-          key={interactionStep.id}
-        >
-          <CardHeader
-            style={styles.cardHeader}
-            title={title}
-            subtitle={
-              interactionStep.parentInteractionId
-                ? ""
-                : "Enter a script for your texter along with the question you want the texter be able to answer on behalf of the contact."
-            }
-          />
-          <CardText>
-            <GSForm
-              {...dataTest(
-                "childInteraction",
-                !interactionStep.parentInteractionId
-              )}
-              schema={this.formSchema}
-              value={{
-                ...interactionStep,
-                ...(interactionStep.answerActionsData && {
-                  answerActionsData:
-                    typeof interactionStep.answerActionsData === "string"
-                      ? JSON.parse(interactionStep.answerActionsData)
-                      : interactionStep.answerActionsData
-                })
-              }}
-              onChange={this.handleFormChange.bind(this)}
-            >
-              {interactionStep.parentInteractionId ? (
-                <Form.Field
-                  {...dataTest("answerOption")}
-                  name="answerOption"
-                  label="Answer"
-                  fullWidth
-                  hintText="Answer to the previous question"
-                />
-              ) : (
-                ""
-              )}
-              {interactionStep.parentInteractionId &&
-              this.props.availableActions &&
-              this.props.availableActions.length ? (
-                <div key={`answeractions-${interactionStep.id}`}>
-                  <div>
-                    <Form.Field
-                      {...dataTest("actionSelect")}
-                      floatingLabelText="Action handler"
-                      name="answerActions"
-                      type="select"
-                      default=""
-                      choices={[
-                        ...this.props.availableActions.map(action => ({
-                          value: action.name,
-                          label: action.displayName
-                        }))
-                      ]}
-                    />
-                    <IconButton tooltip="An action is something that is triggered by this answer being chosen, often in an outside system">
-                      <HelpIconOutline />
-                      <div></div>
-                    </IconButton>
-                    {instructions ? <div>{instructions}</div> : null}
-                  </div>
-                  {clientChoiceData && clientChoiceData.length ? (
-                    <div>
-                      <Form.Field
-                        {...dataTest("actionDataAutoComplete")}
-                        hintText="Start typing to search for the data to use with the answer action"
-                        floatingLabelText="Answer Action Data"
-                        fullWidth
-                        name="answerActionsData"
-                        type="autocomplete"
-                        choices={clientChoiceData.map(item => ({
-                          value: item.details,
-                          label: item.name
-                        }))}
-                      />
-                      {interactionStep.needRequiredAnswerActionsData ? (
-                        <div className={css(styleSheet.errorMessage)}>
-                          Action requires additional data. Please select
-                          something.
-                        </div>
-                      ) : null}
-                    </div>
-                  ) : null}
-                </div>
-              ) : (
-                ""
-              )}
-              <Form.Field
-                {...dataTest("editorInteraction")}
-                name="script"
-                type="script"
-                fullWidth
-                customFields={this.props.customFields}
-                label="Script"
-                multiLine
-                hintText="This is what your texters will send to your contacts. E.g. Hi, {firstName}. It's {texterFirstName} here."
-              />
-              <Form.Field
-                {...dataTest("questionText")}
-                name="questionText"
-                label="Question"
-                fullWidth
-                hintText="A question for texters to answer. E.g. Can this person attend the event?"
-              />
-            </GSForm>
-          </CardText>
-        </Card>
-        <div style={styles.answerContainer}>
-          {interactionStep.questionText &&
-          interactionStep.script &&
-          (!interactionStep.parentInteractionId ||
-            interactionStep.answerOption) ? (
-            <div>
-              <RaisedButton
-                {...dataTest("addResponse")}
-                label="+ Add a response"
-                onTouchTap={this.addStep(interactionStep.id).bind(this)}
-                style={{ marginBottom: "10px" }}
-              />
-            </div>
-          ) : (
-            ""
-          )}
-          {this.state.displayAllSteps &&
-            interactionStep.interactionSteps
-              .filter(is => !is.isDeleted)
-              .map(is => (
-                <div>
-                  {this.renderInteractionStep(
-                    is,
-                    availableActions,
-                    `Question: ${interactionStep.questionText}`
-                  )}
-                </div>
-              ))}
-        </div>
-      </div>
-    );
-  }
-
   render() {
-    const availableActions = this.props.availableActions.reduce(
-      (result, action) => {
-        const toReturn = {
-          ...result
-        };
-        toReturn[action.name] = action;
-        return toReturn;
-      },
-      {}
-    );
-
-    const tree = makeTree(this.state.interactionSteps);
+    const parentStepTree = makeTree(this.state.interactionSteps);
 
     return (
       <div>
@@ -443,7 +186,20 @@ export default class CampaignInteractionStepsForm extends React.Component {
           title="What do you want to discuss?"
           subtitle="You can add scripts and questions and your texters can indicate responses from your contacts. For example, you might want to collect RSVPs to an event or find out whether to follow up about a different volunteer activity."
         />
-        {this.renderInteractionStep(tree, availableActions)}
+        <InteractionStep
+          interactionStep={parentStepTree}
+          availableActions={this.props.availableActions}
+          title="Start"
+          customFields={this.props.customFields}
+          handlers={{
+            addStep: this.addStep.bind(this),
+            deleteStep: this.deleteStep.bind(this),
+            bumpStep: this.bumpStep.bind(this),
+            topStep: this.topStep.bind(this),
+            bottomStep: this.bottomStep.bind(this),
+            onFormChange: this.handleFormChange.bind(this)
+          }}
+        />
         <RaisedButton
           {...dataTest("interactionSubmit")}
           disabled={this.state.interactionSteps.some(
